+++ RouteOptimization/Controls/MapBuilder/IEdge.cs
@@ -8,9 +8,9 @@ namespace RouteOptimization.Controls.MapBuilder
 {
     public interface IEdge
     {
-        double StartX { get; set; }
-        double StartY { get; set; }
-        double FinishX { get; set; }
-        double FinishY { get; set; }
+        float StartX { get; set; }
+        float StartY { get; set; }
+        float FinishX { get; set; }
+        float FinishY { get; set; }
     }
 }
